refactor(lifecycle-manager)!: Take a `Path` over a string
In the lifecycle manager, this changes it so the given argument is a path
rather than a string. While I was updating the function, I also made sure
data was added as a field rather than interpolated in logs. The ID and
path didn't need to be added because the instrument macro adds it to the
root span for the function

Signed-off-by: Taylor Thomas <taylor@oftaylor.com>