docs: Updates README with additional information and add License
Adds some additional information for getting started, and introduction,
and a simple contributing section. I also tried to clean up some flow and
elaborate on some of the other benefits of the component model.

We still have some work to do on making it easy to run the example, so I
left that as a TODO for now

Signed-off-by: Taylor Thomas <taylor@oftaylor.com>