feat(wassette): unload-component now should delete files on Disk
Currently, unload-component API only deletes the runtime state of the Wassette server, but it doesn not touch the files on the Disk. There is a separate API called uninstall-component. However, this is not a symmetric to load-component that users would expect. This commit deletes the unused uninstall-component function and modifies the unload-componnet API to be able to delete the wasm and co-located policy files on Disk.

Signed-off-by: Jiaxiao Zhou <duibao55328@gmail.com>